refactor: hook only
