+++ src/Nfts/NftsNavigator.tsx
@@ -1,44 +1,15 @@
 import {createStackNavigator} from '@react-navigation/stack'
 import React from 'react'
-import {defineMessages, useIntl} from 'react-intl'
 
 import {NftRoutes} from '../navigation'
-import {useSearchHeaderOptions} from '../Search/SearchHeader'
 import {Nfts} from './Nfts'
 
 const Stack = createStackNavigator<NftRoutes>()
 
 export const NftsNavigator = () => {
-  const strings = useStrings()
-  const searchHeaderOptions = useSearchHeaderOptions({
-    placeHolderText: strings.search,
-    title: strings.title,
-    noBack: true,
-  })
-
   return (
     <Stack.Navigator>
-      <Stack.Screen name="nft-gallery" component={Nfts} options={searchHeaderOptions} />
+      <Stack.Screen name="nft-gallery" component={Nfts} />
     </Stack.Navigator>
   )
 }
-
-const useStrings = () => {
-  const intl = useIntl()
-
-  return {
-    title: intl.formatMessage(messages.title),
-    search: intl.formatMessage(messages.search),
-  }
-}
-
-const messages = defineMessages({
-  title: {
-    id: 'nft.navigation.title',
-    defaultMessage: '!!!NFT Gallery',
-  },
-  search: {
-    id: 'nft.navigation.search',
-    defaultMessage: '!!!Search NFT',
-  },
-})
