+++ src/Nfts/NoNftsScreen.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import {Text} from 'react-native'
 
 import {QueryProvider} from '../../.storybook/decorators'
-import {SearchProvider} from '../Search'
+import {SearchProvider} from '../Search/SearchContext'
 import {SelectedWalletProvider} from '../SelectedWallet'
 import {mocks} from '../yoroi-wallets/mocks'
 import {NoNftsScreen} from './NoNftsScreen'
