--- src/Search/index.tsx
@@ -1,2 +1,0 @@
-export {SearchProvider, useSearch} from './SearchContext'
-export {SearchHeader} from './SearchHeader'
