+++ src/features/Send/useCases/ListAmountsToSend/AddToken/SelectTokenFromListScreen.tsx
@@ -1,16 +1,13 @@
 import {useNavigation} from '@react-navigation/native'
 import {FlashList} from '@shopify/flash-list'
 import React from 'react'
-import {defineMessages, useIntl} from 'react-intl'
-import {Image, StyleSheet, TouchableOpacity, View} from 'react-native'
+import {StyleSheet, TouchableOpacity, View} from 'react-native'
 
-import noAssetsImage from '../../../../../assets/img/no-assets-found.png'
 import {Boundary, Spacer, Text} from '../../../../../components'
 import {AmountItem} from '../../../../../components/AmountItem/AmountItem'
-import globalMessages, {txLabels} from '../../../../../i18n/global-messages'
 import {TxHistoryRouteNavigation} from '../../../../../navigation'
-import {useSearch} from '../../../../../Search'
-import {useSelectedWallet} from '../../../../../SelectedWallet'
+import {useSearch, useSearchOnNavBar} from '../../../../../Search/SearchContext'
+import {useSelectedWallet} from '../../../../../SelectedWallet/Context/SelectedWalletContext'
 import {sortTokenInfos} from '../../../../../utils'
 import {YoroiWallet} from '../../../../../yoroi-wallets/cardano/types'
 import {maxTokensPerTx} from '../../../../../yoroi-wallets/contants'
@@ -19,12 +16,20 @@ import {TokenInfo} from '../../../../../yoroi-wallets/types'
 import {Amounts, Quantities} from '../../../../../yoroi-wallets/utils'
 import {filterAssets} from '../../../common/filterAssets'
 import {useSelectedTokensCounter, useSend, useTokenQuantities} from '../../../common/SendContext'
-import {MaxTokensPerTx} from './ShowError/MaxTokensPerTx'
+import {useStrings} from '../../../common/strings'
+import {EmptySearchResult} from './Show/EmptySearchResult'
+import {MaxTokensPerTx} from './Show/MaxTokensPerTx'
 
 export const SelectTokenFromListScreen = () => {
   const strings = useStrings()
-  const wallet = useSelectedWallet()
 
+  // use case: search listed tokens
+  useSearchOnNavBar({
+    placeholder: strings.searchTokens,
+    title: strings.selecteAssetTitle,
+  })
+
+  const wallet = useSelectedWallet()
   const balances = useBalances(wallet)
 
   const tokenInfos = useTokenInfos({
@@ -36,18 +41,17 @@ export const SelectTokenFromListScreen = () => {
 
   const {search: assetSearchTerm} = useSearch()
   const sortedTokenInfos = sortTokenInfos({wallet, tokenInfos: filterAssets(assetSearchTerm, tokenInfos)})
+  const isSearchResultEmpty = assetSearchTerm.length > 0 && sortedTokenInfos.length === 0
 
   return (
-    <View style={styles.container}>
-      <View style={styles.subheader}>
-        {!canAddToken && (
-          <>
-            <MaxTokensPerTx />
+    <View style={styles.root}>
+      {!canAddToken && (
+        <View style={styles.panel}>
+          <MaxTokensPerTx />
 
-            <Spacer height={16} />
-          </>
-        )}
-      </View>
+          <Spacer height={16} />
+        </View>
+      )}
 
       <FlashList
         data={sortedTokenInfos}
@@ -61,7 +65,7 @@ export const SelectTokenFromListScreen = () => {
         keyExtractor={(_, index) => index.toString()}
         testID="assetsList"
         estimatedItemSize={78}
-        ListEmptyComponent={assetSearchTerm.length > 0 && sortedTokenInfos.length === 0 ? <NoAssets /> : undefined}
+        ListEmptyComponent={isSearchResultEmpty ? <EmptySearchResult /> : undefined}
       />
 
       <View style={styles.counter}>
@@ -98,52 +102,6 @@ const SelectableAssetItem = ({tokenInfo, disabled, wallet}: SelectableAssetItemP
   )
 }
 
-export const NoAssets = () => {
-  const strings = useStrings()
-  return (
-    <View style={styles.imageContainer}>
-      <Spacer height={160} />
-
-      <Image source={noAssetsImage} style={styles.image} />
-
-      <Spacer height={25} />
-
-      <Text style={styles.contentText}>{strings.noAssets}</Text>
-    </View>
-  )
-}
-const useStrings = () => {
-  const intl = useIntl()
-
-  return {
-    unknownAsset: intl.formatMessage(messages.unknownAsset),
-    assetsLabel: intl.formatMessage(globalMessages.assetsLabel),
-    amount: intl.formatMessage(txLabels.amount),
-    counter1: (count) => intl.formatMessage(messages.counter1, {count}),
-    counter2: intl.formatMessage(messages.counter2),
-    noAssets: intl.formatMessage(messages.noAssets),
-  }
-}
-
-const messages = defineMessages({
-  unknownAsset: {
-    id: 'components.send.assetselectorscreen.unknownAsset',
-    defaultMessage: '!!!Unknown asset',
-  },
-  counter1: {
-    id: 'components.send.assetselectorscreen.counter1',
-    defaultMessage: '!!!{count} assets',
-  },
-  counter2: {
-    id: 'components.send.assetselectorscreen.counter2',
-    defaultMessage: '!!!found',
-  },
-  noAssets: {
-    id: 'components.send.assetselectorscreen.noAssets',
-    defaultMessage: '!!!No assets found',
-  },
-})
-
 const styles = StyleSheet.create({
   counter: {
     justifyContent: 'center',
@@ -160,29 +118,12 @@ const styles = StyleSheet.create({
   item: {
     paddingVertical: 16,
   },
-  subheader: {
+  panel: {
     paddingTop: 16,
     paddingHorizontal: 16,
   },
-  container: {
+  root: {
     flex: 1,
     backgroundColor: 'white',
-  },
-  contentText: {
-    flex: 1,
-    textAlign: 'center',
-    fontWeight: '700',
-    fontSize: 20,
-    color: '#000',
-  },
-  image: {
-    flex: 1,
-    alignSelf: 'center',
-    width: 200,
-    height: 228,
-  },
-  imageContainer: {
-    flex: 1,
-    textAlign: 'center',
   },
 })
