+++ src/i18n/locales/en-US.json
@@ -182,14 +182,13 @@
   "components.send.sendscreen.feeLabel": "Fee",
   "components.send.sendscreen.feeNotAvailable": "-",
   "components.send.sendscreen.resolvesTo": "Resolves to",
+  "components.send.sendscreen.searchTokens": "Search assets",
   "components.send.sendscreen.sendAllWarningText": "You have selected the send all option. Please confirm that you understand how this feature works.",
   "components.send.sendscreen.sendAllWarningTitle": "Do you really want to send all?",
   "components.send.sendscreen.sendAllWarningAlert1": "All your {assetNameOrId} balance will be transferred in this transaction.",
   "components.send.sendscreen.sendAllWarningAlert2": "All your tokens, including NFTs and any other native assets in your wallet, will also be transferred in this transaction.",
   "components.send.sendscreen.sendAllWarningAlert3": "After you confirm the transaction in the next screen, your wallet will be emptied.",
   "components.send.sendscreen.title": "Send",
-  "components.send.sendscreen.searchTitle": "Search assets",
-  "components.send.sendscreen.searchPlaceholder": "Search assets",
   "components.settings.applicationsettingsscreen.biometricsSignIn": "Sign in with your biometrics",
   "components.settings.applicationsettingsscreen.changePin": "Change PIN",
   "components.settings.applicationsettingsscreen.commit": "Commit:",
